backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9